requirements.txt
requests~=2.32.3
playsound~=1.3.0
python-dotenv~=1.0.1
pylint~=3.3.2
